+++ src/app/events/MessageCreate.ts
@@ -1,5 +1,4 @@
 import messageCreateOnBountyBoard from './bounty/MessageCreateOnBountyBoard';
-import messageSetScoapRoles from './scoap-squad/messageSetScoapRoles';
 import { Message } from 'discord.js';
 import { DiscordEvent } from '../types/discord/DiscordEvent';
 import MessageCreateOnDEGEN from './chat/MessageCreateOnDEGEN';
@@ -21,12 +20,13 @@ export default class implements DiscordEvent {
 			await messageCreateOnBountyBoard(message).catch(e => {
 				LogUtils.logError('failed to create bounty message from webhook', e);
 			});
-			if (message.channel.type === 'DM') {
-				// Run scoap squad DM flow
-				await messageSetScoapRoles(message).catch(e => {
-					LogUtils.logError('failed to run scoap-squad DM flow', e);
-				});
-			}
+			// TODO: enable scoap squad at a future date
+			// if (message.channel.type === 'DM') {
+			// 	// Run scoap squad DM flow
+			// 	await messageSetScoapRoles(message).catch(e => {
+			// 		LogUtils.logError('failed to run scoap-squad DM flow', e);
+			// 	});
+			// }
 			// Check mentions for AFK users
 			if (message.mentions.users.size > 0) {
 				await HandleAFK(message).catch((e) => {